Add files via upload
This code was developed by Marco Baldo and was needed to synthesize and harmonize the Bottoms-Up database in the study region of the Czech Republic [Biodivers Conserv (2015) 24:1391-1402] with the process-based forest dynamics model iLand [Seidl et al., 2012]. The database consists of 99 plots of 2500 sq. meters (50x50) of forest structures and necromass, multi-taxonomic samples, and environmental variables (climate, soil structure, topography, etc).